backend/requirements.txt
# Django & DRF & JWT
django==5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
# Control Cors headers
django_cors_headers==4.4.0
# Postgres client and dependencies
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
# Load dotenv files
python-dotenv==1.0.1
django-environ==0.11.2
# uvicorn server
uvicorn==0.30.6
# Colorize test outputs
redgreenunittest==0.1.1
# Django redis client
django-redis==5.4.0
# Celery
celery==5.4.0
redis==5.0.8
amqp==5.2.0
msgpack==1.0.8
django-celery-results==2.5.1
django-celery-beat==2.7.0
flower==2.0.1
# SMSProvider
kavenegar==1.1.2
# Admin UI
django-jazzmin==3.0.0
# Formatting
ruff==0.6.4
mypy==1.11.2
# Swagger
drf-yasg==1.21.7
# Load testing
locust==2.31.6
# Test coverage
coverage==7.6.1
